+++ CHANGELOG.md
@@ -9,6 +9,8 @@ volumes to mount into Docker containers created by plugins.
 
 ### Changed
 
+- ansible: fixed handling of `auth` parameters. The deprecated `user`
+and `privateKey` parameters are working now again;
 - concord-console: the URL parsing in the log viewer is updated to
 better handle URLs in quotes.